more splits, renames and cleanup - no functional changes
