+++ src/daemon/pulse/pulse-db-dbengine.c
@@ -1490,7 +1490,7 @@ void pulse_dbengine_do(bool extended) {
 
                 /* get localhost's DB engine's statistics for each tier */
                 for(size_t tier = 0; tier < nd_profile.storage_tiers;tier++) {
-                    if(host->db[tier].mode != RRD_MEMORY_MODE_DBENGINE) continue;
+                    if(host->db[tier].mode != RRD_DB_MODE_DBENGINE) continue;
                     if(!host->db[tier].si) continue;
 
                     if(counted_multihost_db[tier])
