+++ src/database/rrd-algorithm.c
@@ -1,0 +1,37 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#include "rrd-algorithm.h"
+
+RRD_ALGORITHM rrd_algorithm_id(const char *name) {
+    if(strcmp(name, RRD_ALGORITHM_INCREMENTAL_NAME) == 0)
+        return RRD_ALGORITHM_INCREMENTAL;
+
+    else if(strcmp(name, RRD_ALGORITHM_ABSOLUTE_NAME) == 0)
+        return RRD_ALGORITHM_ABSOLUTE;
+
+    else if(strcmp(name, RRD_ALGORITHM_PCENT_OVER_ROW_TOTAL_NAME) == 0)
+        return RRD_ALGORITHM_PCENT_OVER_ROW_TOTAL;
+
+    else if(strcmp(name, RRD_ALGORITHM_PCENT_OVER_DIFF_TOTAL_NAME) == 0)
+        return RRD_ALGORITHM_PCENT_OVER_DIFF_TOTAL;
+
+    else
+        return RRD_ALGORITHM_ABSOLUTE;
+}
+
+const char *rrd_algorithm_name(RRD_ALGORITHM algorithm) {
+    switch(algorithm) {
+        case RRD_ALGORITHM_ABSOLUTE:
+        default:
+            return RRD_ALGORITHM_ABSOLUTE_NAME;
+
+        case RRD_ALGORITHM_INCREMENTAL:
+            return RRD_ALGORITHM_INCREMENTAL_NAME;
+
+        case RRD_ALGORITHM_PCENT_OVER_ROW_TOTAL:
+            return RRD_ALGORITHM_PCENT_OVER_ROW_TOTAL_NAME;
+
+        case RRD_ALGORITHM_PCENT_OVER_DIFF_TOTAL:
+            return RRD_ALGORITHM_PCENT_OVER_DIFF_TOTAL_NAME;
+    }
+}
