+++ src/database/rrdhost-labels.h
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+#ifndef NETDATA_RRDHOST_LABELS_H
+#define NETDATA_RRDHOST_LABELS_H
+
+#include "libnetdata/libnetdata.h"
+
+void reload_host_labels(void);
+void rrdhost_set_is_parent_label(void);
+
+#endif //NETDATA_RRDHOST_LABELS_H
