+++ src/database/rrdset-collection.c
@@ -273,83 +273,6 @@ static inline void rrdset_init_last_updated_time(RRDSET *st) {
 
 __thread size_t rrdset_done_statistics_points_stored_per_tier[RRD_STORAGE_TIERS];
 
-#ifdef NETDATA_LOG_COLLECTION_ERRORS
-void rrddim_store_metric_with_trace(RRDDIM *rd, usec_t point_end_time_ut, NETDATA_DOUBLE n, SN_FLAGS flags, const char *function) {
-#else // !NETDATA_LOG_COLLECTION_ERRORS
-void rrddim_store_metric(RRDDIM *rd, usec_t point_end_time_ut, NETDATA_DOUBLE n, SN_FLAGS flags) {
-#endif // !NETDATA_LOG_COLLECTION_ERRORS
-
-    static __thread struct log_stack_entry lgs[] = {
-        [0] = ND_LOG_FIELD_STR(NDF_NIDL_DIMENSION, NULL),
-        [1] = ND_LOG_FIELD_END(),
-    };
-    lgs[0].str = rd->id;
-    log_stack_push(lgs);
-
-#ifdef NETDATA_LOG_COLLECTION_ERRORS
-    rd->rrddim_store_metric_count++;
-
-    if(likely(rd->rrddim_store_metric_count > 1)) {
-        usec_t expected = rd->rrddim_store_metric_last_ut + rd->update_every * USEC_PER_SEC;
-
-        if(point_end_time_ut != rd->rrddim_store_metric_last_ut) {
-            internal_error(true,
-                           "%s COLLECTION: 'host:%s/chart:%s/dim:%s' granularity %d, collection %zu, expected to store at tier 0 a value at %llu, but it gave %llu [%s%llu usec] (called from %s(), previously by %s())",
-                           (point_end_time_ut < rd->rrddim_store_metric_last_ut) ? "**PAST**" : "GAP",
-                           rrdhost_hostname(rd->rrdset->rrdhost), rrdset_id(rd->rrdset), rrddim_id(rd),
-                           rd->update_every,
-                           rd->rrddim_store_metric_count,
-                           expected, point_end_time_ut,
-                           (point_end_time_ut < rd->rrddim_store_metric_last_ut)?"by -" : "gap ",
-                           expected - point_end_time_ut,
-                           function,
-                           rd->rrddim_store_metric_last_caller?rd->rrddim_store_metric_last_caller:"none");
-        }
-    }
-
-    rd->rrddim_store_metric_last_ut = point_end_time_ut;
-    rd->rrddim_store_metric_last_caller = function;
-#endif // NETDATA_LOG_COLLECTION_ERRORS
-
-    // store the metric on tier 0
-    storage_engine_store_metric(rd->tiers[0].sch, point_end_time_ut,
-                                n, 0, 0,
-                                1, 0, flags);
-
-    rrdset_done_statistics_points_stored_per_tier[0]++;
-
-    time_t now_s = (time_t)(point_end_time_ut / USEC_PER_SEC);
-
-    STORAGE_POINT sp = {
-        .start_time_s = now_s - rd->rrdset->update_every,
-        .end_time_s = now_s,
-        .min = n,
-        .max = n,
-        .sum = n,
-        .count = 1,
-        .anomaly_count = (flags & SN_FLAG_NOT_ANOMALOUS) ? 0 : 1,
-        .flags = flags
-    };
-
-    for(size_t tier = 1; tier < nd_profile.storage_tiers;tier++) {
-        if(unlikely(!rd->tiers[tier].smh)) continue;
-
-        struct rrddim_tier *t = &rd->tiers[tier];
-
-        if(!rrddim_option_check(rd, RRDDIM_OPTION_BACKFILLED_HIGH_TIERS)) {
-            // we have not collected this tier before
-            // let's fill any gap that may exist
-            backfill_tier_from_smaller_tiers(rd, tier, now_s);
-        }
-
-        store_metric_at_tier(rd, tier, t, sp, point_end_time_ut);
-    }
-    rrddim_option_set(rd, RRDDIM_OPTION_BACKFILLED_HIGH_TIERS);
-
-    rrdcontext_collected_rrddim(rd);
-    log_stack_pop(&lgs);
-}
-
 // caching of dimensions rrdset_done() and rrdset_done_interpolate() loop through
 struct rda_item {
     const DICTIONARY_ITEM *item;
@@ -360,7 +283,7 @@ struct rda_item {
 static __thread struct rda_item *thread_rda = NULL;
 static __thread size_t thread_rda_entries = 0;
 
-struct rda_item *rrdset_thread_rda_get(size_t *dimensions) {
+static struct rda_item *rrdset_thread_rda_get(size_t *dimensions) {
 
     if(unlikely(!thread_rda || (*dimensions) > thread_rda_entries)) {
         size_t old_mem = thread_rda_entries * sizeof(struct rda_item);
@@ -665,7 +588,7 @@ void rrdset_timed_done(RRDSET *st, struct timeval now, bool pending_rrdset_next)
 
     // check if we will re-write the entire data set
     if(unlikely(dt_usec(&st->last_collected_time, &st->last_updated) > st->db.entries * update_every_ut &&
-                 st->rrd_memory_mode != RRD_MEMORY_MODE_DBENGINE)) {
+                 st->rrd_memory_mode != RRD_DB_MODE_DBENGINE)) {
         nd_log_daemon(NDLP_DEBUG, "'%s': too old data (last updated at %" PRId64 ".%" PRId64 ", last collected at %" PRId64 ".%" PRId64 "). "
                                   "Resetting it. Will not store the next entry.",
                       rrdset_id(st),
