+++ src/streaming/stream-conf.h
@@ -52,7 +52,7 @@ struct _stream_receive {
 extern struct _stream_receive stream_receive;
 
 struct stream_receiver_config {
-    RRD_MEMORY_MODE mode;
+    RRD_DB_MODE mode;
     bool ephemeral;
     int history;
     int update_every;
