+++ src/streaming/stream-parents.h
@@ -17,16 +17,10 @@ typedef struct rrdhost_stream_parents {
 } RRDHOST_STREAM_PARENTS;
 
 #include "stream-handshake.h"
-#include "database/rrdhost-status.h"
 
 void rrdhost_stream_parent_ssl_init(struct sender_state *s);
 
 int stream_info_to_json_v1(BUFFER *wb, const char *machine_guid);
-
-void rrdhost_stream_parents_reset(RRDHOST *host, STREAM_HANDSHAKE reason);
-
-void rrdhost_stream_parents_update_from_destination(RRDHOST *host);
-void rrdhost_stream_parents_free(struct rrdhost *host, bool having_write_lock);
 
 bool stream_parent_connect_to_one(
     ND_SOCK *sender_sock,
@@ -46,6 +40,11 @@ bool stream_parent_is_ssl(STREAM_PARENT *d);
 
 usec_t stream_parent_handshake_error_to_json(BUFFER *wb, struct rrdhost *host);
 
-void rrdhost_stream_parents_init(RRDHOST *host);
+void rrdhost_stream_parents_reset(struct rrdhost *host, STREAM_HANDSHAKE reason);
+
+void rrdhost_stream_parents_update_from_destination(struct rrdhost *host);
+void rrdhost_stream_parents_free(struct rrdhost *host, bool having_write_lock);
+
+void rrdhost_stream_parents_init(struct rrdhost *host);
 
 #endif //NETDATA_STREAM_PARENTS_H
